Metrics Reporting Provisioning API Refactored
