+++ src/5gmsaf/provisioning-session.h
@@ -46,21 +46,6 @@ typedef struct msaf_provisioning_session_s {
 
 } msaf_provisioning_session_t;
 
-typedef struct msaf_metrics_reporting_configuration_s {
-    char *metricsReportingConfigurationId;
-    char *scheme;
-    char *dataNetworkName;
-    bool isReportingInterval;
-    int reportingInterval;
-    bool isSamplePercentage;
-    double samplePercentage;
-    OpenAPI_list_t *urlFilters;
-    OpenAPI_list_t *metrics;
-    //char *provisioningSessionId;
-    char *etag;
-    time_t receivedTime;
-    char *metricsReportingConfigurationHash;
-} msaf_metrics_reporting_configuration_t;
 
 typedef struct msaf_application_server_state_node_s msaf_application_server_state_node_t;
 typedef struct msaf_application_server_state_ref_node_s {
@@ -72,40 +57,12 @@ extern msaf_provisioning_session_t *msaf_provisioning_session_create(const char 
 extern msaf_provisioning_session_t *msaf_provisioning_session_find_by_provisioningSessionId(const char *provisioningSessionId);
 extern cJSON *msaf_provisioning_session_get_json(const char *provisioning_session_id);
 
-extern msaf_metrics_reporting_configuration_t *msaf_metrics_reporting_configuration_create(msaf_provisioning_session_t *provisioning_session,
-                                                                                           const char *metricsReportingConfigurationId,
-                                                                                           const char *scheme,
-                                                                                           const char *data_network_name,
-                                                                                           bool is_reporting_interval,
-                                                                                           int reporting_interval,
-                                                                                           bool is_sample_percentage,
-                                                                                           double sample_percentage,
-                                                                                           OpenAPI_list_t *url_filters,
-                                                                                           OpenAPI_list_t *metrics);
-
-extern msaf_metrics_reporting_configuration_t* mrc_update(const char *metricsReportingConfigurationId,
-                                                          const char *scheme,
-                                                          const char *dataNetworkName,
-                                                          bool isReportingInterval,
-                                                          int reportingInterval,
-                                                          bool isSamplePercentage,
-                                                          double samplePercentage,
-                                                          OpenAPI_list_t *urlFilters,
-                                                          OpenAPI_list_t *metrics);
-
-extern msaf_metrics_reporting_configuration_t* mrc_retrieve(const char *metricsReportingConfigurationId);
-extern cJSON *msaf_metrics_reporting_configuration_get_json(const char *metrics_reporting_configuration_id);
-extern int msaf_metrics_reporting_configuration_delete(const char *metricsReportingConfigurationId);
-extern void msaf_provisioning_session_mrc_hash_remove(const char *provisioning_session_id, const char *metricsReportingConfigurationId);
-extern int msaf_metrics_reporting_configuration_delete(const char *metricsReportingConfigurationId);
-
 extern OpenAPI_content_hosting_configuration_t *msaf_content_hosting_configuration_create(msaf_provisioning_session_t *provisioning_session);
 
 extern int msaf_content_hosting_configuration_certificate_check(msaf_provisioning_session_t *provisioning_session);
 extern int msaf_distribution_certificate_check(void);
 
 extern ogs_hash_t *msaf_certificate_map();
-extern ogs_hash_t *msaf_metrics_reporting_map();
 extern const char *msaf_get_certificate_filename(const char *provisioning_session_id, const char *certificate_id);
 extern ogs_list_t *msaf_retrieve_certificates_from_map(msaf_provisioning_session_t *provisioning_session);
